data-pipeline/requirements.txt
pymongo
boto3
pyspark
apache-airflow
apache-airflow-providers-apache-spark 
pyarrow
s3fs
fsspec